Move `pycors()` to `main.rs`
